Planet sprite
This adds a sprite for the planet, since it represents an extremely
distant object, we don't want distortion as it moves to the edge of the
screen.

The picture used is a Public Domain one from Renderwahn, and originally
meant for NAEV. I think they use much larger sprites now. Note to self:
See if they made bigger new planet sprites.