Adds timer to EntMan
This adds a timer to EntMan, which allows systems to access
the time delta with entMan.deltaTime. It uses the default
javascript Date.now() - it might be possible to do this in
a nicer way with babalon js in the future.

It also might make sense to factor the timer out into a
service so that it does not need to clutter up the entMan,
since it's got nothing to do with the entities.