Fullscreen map
The map view now reads the current size of the canvas and uses that to
calculate where to center the map. This offset is saved as diff and the
difference in offset is returned when the map is closed so that it can
be returned when the map is resized.