Replaces "Eich" with "Reese"
This was a tough one.

Eich was, obviously, named for the creator of Javascript, Brandon Eich.
The commit was written on August 5th, 2013. The last development on that
project was in March 3rd. Eich was appointed as the CEO of Mozilla on
the 24th and this set off a firestorm. This is a risk you take when you
name something after a clearly identifiable living person. I feel that
keeping the name there is making a statement that I don't want to make.
I've replaced it with a different word - Reese.