Close map
This enables the map to close with the new state system. This required
two changes: The GameplayState needs to clean up its models and entites
at the right moments, and only re-create them if they're gone.

The other, (rather surprising) issue was that when I left the map state,
it was hitting 'escape' a second time and re-opening the map. The
solution to this was to debounce the input on that key.

The input system could probably use a refactor. It might make sense to
add a bunch of input functions to the state class. It might make sense
to add a class that does the debouncing.