Implements land / takeoff
Realized that I'd already factored out the menu system hiding inside the
map state, so I got rid of LandingView in favor of MainMenuView (which
should probably actually be MenuView) and implemented LandingMainView
off of that.

Probably going to need to build a primative GUI system since this is a
GUI heavy game that I'm building.