Refactors menu views to use a widget class
The OOP train has no brakes. The LandingMainView is now just a
constructor which adds a 'Leave' widget (specifically a text button,
which takes text to display and a function to call when the button is
pressed, a location, and a label.

The one flaw in this current setup is that you need to both pass the
id and specify it as a key to widgets - that's sort of gross because
if you mistype the key either time it won't work and there should be
a single source of truth for that sort of information.