Moves some init from load into main
In preparation for using the babylon asset manager, this moves the
engine and scene declarations into load.js, so that the asset manager
can use them. It's also got a shell function 'load_assets' which is
called when we finish loading 'assets.json'. Now, I'm not sure if it's
possible to load json with the asset loader, but if it is, it would make
sense to use it for all of the asset loading, rather than jq