Restructure code around AssetsManager
This restructures the code in load around AssetsManager rather than a
bunch of jquery calls. This moves is closer to removing jq as a
dependancy and possibly moving this code to main. Note that there is
some commented out code in this commit. The first chunk loads models,
and the next chunk is supposed to load text but does not work. The
problem with loading models with the AssetsManager is that it attaches
them direcly to the scene - I'd prefer caching them then instancing them
at a later time. I'll have to look into if that's possible.