Use loader for text
This replaces the jquery json loading with the babylon AssetsManager so
that I can wait until everything is loaded without a bunch of promise
logic. This also makes it possible to add new text files very easily. In
order to make it simpler, I've broken up the data files such that every
type dict is in its own file. I'm not sure why everything on the map was
in its own file besides the simplicity it brought to a notepad++
oriented workflow.