Venerable AI code
This AI code was pasted in from Eridanus, but that's not really the full
story. It was actually the last part of Erndaius that I worked on, the
last big push on that project before I ultimately dropped it. But it was
just light modifications of the original AI code, which came from my
"asteroids" proof of concept demo. In that demo I proved to myself that
I could do sprites, collisions, and AI. By my reasoning at the time,
that meant that I could easily tackle the rest of a 2d EV clone. I was
not correct, but I did learn something important - do the hard stuff
first. I've tried to do the same in this project.