AI is able to chase down and destroy the asteroid
The trick turned out to be imposing some sort of reasonable speed limit.
Still have some lingering doubts that it is firing projectiles in the
right direction. Will need to rig up a foe ship to test further. This is
as good as it is going to get with the AI superimposed over player
control.