NPCs won't hurt themselves or allies
As improbable as it is, our ships are now firing "smart" projectiles
which will not harm allies. This is a nececity for 2d gameplay and
players will prefer not to see everything blowing itself up.

Might not be crazy to add a "friendly-fire" flag to weapons so you can,
for example, make an AOE missile that you need to be careful with.