Lines drag with map
This refactors how lines are created by moving it out from the main
system create loop and into a seperate one that checks to ensure that
there are no duplicate lines. It also leaves each line in an array so
that we can update them en masse whenever the map is dragged around.