Upgrades landing menu
This upgrades the landing/menu system (widgets and all) from the old
canvas2d style to the new ADT style. This is actually what prompted me to upgrade
the engine - I could not get images to display inside the canvas2d, and
figured this would probably be fixed in a newer version of the engine. What I got
instead was not having to roll so much of my own code to get the GUI working -
widgets now chiefly serve as data to instantiate controls on an ADT. Overall,
I'm very pleased with the way it turned out vs the old system, so right on
babylonjs