This look has a bit of history...
I'm really mining old assets here. This one I made for what was
perhaps one of the earliest incarnations of the project - it was
going to be written in Blitz Basic. I got about as far as
implementing... particle effects. That was it. It ended up being
a pretty cool fireworks simulator. It was called 'Turfless' I think.