Fixes issue where camera would get wonky on liftoff
This is even simpler, and requires no initial setup (which I'm
lothe to do in a system which is supposed to be stateless.) Just
orient the camera towards the origin when the camera is created,
and slide the camera to follow the entity we care about by adding
a fixed camera offset to the entity's position. This seems to
solve the jitter issues.