Randomized NPCs
This adds in a system for procedurally populating the NPCs. It is
broadly similar to the "pers" resource in EV - you define a group
(in npc_groups.json) of possible ships and what government they
should spawn for. Then in the NPCs field in System, you list what
groups can be picked from.