Object.create refactor of weapons
Weapons where previously a class. This lead to an ugly constructor and
got more complex every time I wanted to add a new attribute. This refactor
brings it in line with other objects spawned directly from data, reduces
the amount of work required to add additional weapon qualities, and takes
the oppertunity to update some terminology.