Adds provision for mesh metadata
Trying to embed everything in mesh bone names was proving impossible.
This bundles some metadata with the mesh in assets.json, including what
gun slots belong to which turret. This should allow for turrets with
multiple guns mounted on them. If it where possible to hang arbitrary
data off of bones like you can in blender this wouldn't be needed, but
such data is stripped out by the exporter.