Adds tech filter to spobs
So part of the fun of having different factions with different items
is barring you from buying stuff on the wrong world. This adds a tech
filter to the upgrade store and the shipyard which means only spobs
that meet or exceed the tech levels of items will be able to buy and
sell those items. Also lays some of the groundwork for selling upgrades.