Radically different approach to missions
So dumping a mission's variables into the local scope turned out to
be a bad idea. A better idea was to make a class and implement eveything
as instance methods. But I really didn't want to loose out on the magic
of 'object.create' which has made the entire project less error-prone.

So really I want to multi-inherit from both the data object and the
mission class. Turns out someone on StackOverflow figured out how:

https://stackoverflow.com/a/31236132/1048464

So that's that.