Fifth fleet removed
Just like that.
They weren't really serving a purpose and where just going to be a bunch
of reskins of existing ships. We have enough foes in northern space thanks
to the Loyal Suns, so I'm just going to leave that corner blank for now.

Also, reduicing the number of Warsaw Pact names in Itaskan space in favor
of a more varied set. Mostly because we have more than enough elsewhere
and it's important to distinguish between them and the nearby Baikonour.