Basic cargo mission modals work
Note: if the number of interpolated fields becomes unweildy (and it
quickly is) it might make sense to have either an enumerated list of
interpolated fields, or some such.

Also, note that we're interpolating the modal text at bake time,
but it would be more correct to interpolate upon actually showing
the modal in case something changed.