Adds first unit tests
Developing the line collider code is going to be impossible
without tests. Also, I frequently hear that code is only as
good as the tests that run on it. So it's time to add basic
unit tests to FTS. Because it's FTS, they're going to be
idiosyncratic though. They'll run only if you pass ?test=1
as a querystring. I feel strongly that this whole project
should be browser based so as not to require any setup.