Brakethrough: passing data to the shader
With some help from the fine folks on the BabylonJS forum,
I'm piecing together what it takes to make shaders work.
https://forum.babylonjs.com/t/is-there-a-full-list-of-variables-available-in-shaders-anywhere/7051/4
Looks like to get the 'time' float in, you need to declare
the variable as 'uniform' in the shader, add it to the
'uniforms' array, and set it with setFloat. Time to
experiment a bit more!