Strictly layer on-screen drawing
Just discovered renderingGroupId. See:
https://doc.babylonjs.com/resources/transparency_and_how_meshes_are_rendered

Took the opportunity to centralize graphics a bit more in the process.

The door is still open to draw the player's ship on top, but I don't want
it to get drawn over, say, explosions so I'll need to figure that out.

One approach would be to make the bottom half of the explosion
drawn on a layer below and the top half on a layer above... it
should be possible to restrict particle emission direction to
get such an effect. But the juice might not be worth the squeeze
as it where.