Switch between perspectives - first take
Replaced the camera constant with a getter.

The problem with this approach is that it's very relevant what position
you set the camera to initially. It may make more sense to create a new
camera on game start, or at least move it to the new CAM_OFFSET when the
game play state begins.