Keep track of sprite managers to list star sprites
We want to keep track of the star sprite managers, because we want to
make sure we can enumerate all of the stars in a system. This adds some
tracking (taking advantage of return destructuring) and a method to
flatten the list of stars so that they can be operated on. In this case
the operation looks to be "stretch them out to resemble star trek/ star
wars."