Per-material glows
Due to some help in this thread:
https://forum.babylonjs.com/t/emissivetexture-vs-emissivecolor/7931/5
pointing to this documentation:
https://doc.babylonjs.com/how_to/glow_layer#controlling-glow-color-per-mesh

This sets up a per-mesh glow color selector. It hangs the appropriate
color on the material as 'custom_emissive_color' (an array) and then
rest-operators it in the customEmissiveColorSelector callback.

This is the most complex graphics set up to date, so I decided to break
graphics set up into its own function in graphics.js, along with camera
init. This introduces a bug that puts you in 3dverse mode when you
update settings; will address this in a subsequent change.

Finally, having glowing textures is an opportunity to visually show
when ships have been disabled.