Clean up save/load
Pulls the pasta out of saves.js (which is mostly about being a menu) and
moves the details of saving and loading json files into file_utils.js.
Also cleans up the files themselves quite a bit, partly just to keep
them in the same coding style as the rest of the project.