Remodels Melkin ships; textures Vera
Melkin ships are going to share a lot of components, so I remodeled them
all in one go. Only did the texturing for the Vera in this one. Learned
some things:

* In UV edit mode, _you can edit the UV layout_
* In texture paint mode, _you can mask_

Now I just need a workflow for making seprate layers and I'll be all
set. Maybe I could do that with nodes?