Empfi textures
This is supposed to be the most commonly used fighter in the galaxy,
so I made skins for a bunch of possible factions. Many of these aren't
used yet, but when I do ship variants and such I think it'd be cool to
have some versions of some ships carry second line retrofitted OE
fighters.