Pick up debris from asteroids
Added tiny asteroids which can be picked up. Right now every ship can pick them up,
but you could set that to 'false' in the prototype and add an upgrade that gives it
back. You could also make debris that ships drop when they explode - just use the
'drops' property of the ship.'