Refactor to use a distinct notion of a ship save
In order to make fleets work, we need multiple player ships.

This refactor pulls the notion of a ship - its type, its
upgrades, and its cargo, onto a "ShipSave" class. Stuff relating
to the player's own ship is now in a shipsave called flagship.

Also adds a bunch of tests.